login: use default widget for logging on enter press
Expose the login button in login.ui to retrieve it by the default_widget function in login.
Set the login  button as the default widget when the visible child in main_stack is login,
and connect the function to the visible_child change signal to update the default widget.

fix: https://gitlab.gnome.org/GNOME/fractal/-/issues/788