Add login requests to the sidebar and remove `ContentType`
The `ContentType` could be removed because the needed information is
already contained in the `selected-item`. This also addes `EntryType` to
differentiat between different `Entry`s even tought we currently have
only "Explore". This also cleans up how the selected-item is passed
between `Content` and `Sidebar`

This also replaces the `ToDeviceHandler` with `VerificationList` that's
used to track verifications.