meson: Drop build shell script
This drops the script we used to build fractal. And uses know only
meson. This also drops the need for specifying sources since cargo does
that for us.

This also bumps the meson version.
This also drops the build option `skip-cargo-build`.

Most of the changes are comming from https://gitlab.gnome.org/World/pika-backup